requirements.txt
fastapi
streamlit
openai
python-dotenv 